GigaMap: entity ids are derived with long arithmetic, no longer wrapping at 2^31 (internal#133)
Several places derived an entity id from a segment index with int arithmetic and widened to long only on assignment or at the call, although a GigaMap may hold up to 2^50 entities. All of them broke at the same boundary, entity id 2^31.

GigaMap.Default#iterateIndexed and its two private iterate helpers reported wrapped, negative ids, corrupting every consumer of the id: reindex(), bitmap index registration back-fill, the Lucene back-fill (which persists the id into ENTITY_ID_FIELD) and the jvector graph rebuild. Beyond the plain overflow, level1Exp + level2Exp may legally reach 40 (each caps at 20, only their sum with level3Max is limited to 50) while an int shift uses only the low 5 bits of its distance (JLS 15.19), so i<<32 evaluated to i<<0 == i and high level3 segments reported ids colliding with those of level3 segment 0 rather than merely wrapping.

AbstractBitmapIterating and ResultIdIterator computed the level2 base id as level3Index * LEVEL_2_ID_COUNT, so every query silently returned nothing for entities above 2^31 - independently of iterateIndexed, and even for an index the ordinary add path had filled correctly. ResultIdIterator#calculateIndexBound additionally narrowed the id-range remainder through XTypes.to_int, which throws for maps beyond 2^31; the remainder is now kept as a long and narrowed only in the branch where it is provably below one segment's id count.

ThreadedIterator had the same defect in its valueBaseId computation. Its calculateLevel1SegmentCount also narrowed the segment count, which stays an int because it doubles as an index bound into the off-heap id lists registry: that caps the iterator at roughly 2^43 ids, so the count beyond that is now rejected with an explicit message instead of allocating from a negative length. The empty map needs its own branch there: highestUsedId is nextFreeId() - 1 and hence -1, whose unsigned shift is 2^52 - 1, and the former int truncation happened to land on exactly the 0 the iterator needs to report "no data" - so that case worked purely by accident of the overflow.

Two comments in ThreadedIterator were corrected while reading it. currentRegistryIndex points to a registry segment covering REGISTRY_SEGMENT_SIZE level1 segments, not to a single level1 segment as its comment claimed - which is what makes the REGISTRY_SEGMENT_ID_COUNT multiplier in valueBaseId correct. And hasMoreData compares that index against level1SegmentCount rather than registrySegmentCount, which is not the index's bound but an "is the exhaustion sentinel set?" test; it is correct, and now says so.

That same confusion had sized the registry itself: calculateIdListsRegistryLength allocated one slot per level1 segment, while the worker threads, scrollToNextRegistrySegment and deallocateSegments all bound themselves by registrySegmentCount, so up to 63 of every 64 slots were never touched (74 slots allocated for 2 used at 300k entities). The registry is now sized by registrySegmentCount, which required computing that count before the allocation. The length is 0 for an empty map exactly as before, since ceil(n/64) is 0 only for n == 0.

That zero-slot registry is a zero-length allocation, and XMemory.allocateCleared(0) yields the address 0, on which ThreadLogic's bounds arithmetic fails - a not-positive start address, and a registrySegmentCount - 1 of -1 - so every worker thread of a query over an empty map died with a NumberRangeException. It was invisible: the failures happened on the worker threads, and the reader never consults the registry in that case (hasMoreData() is false from the start), so the query still produced the correct empty result. No worker threads are started when there is no registry segment to process.

The reported ids and their downstream effects are covered by EntityIdAboveIntegerRangeTest, the ThreadedIterator segment count edges, registry sizing and worker thread health by ThreadedIteratorSegmentCountTest - that one deliberately spans more than one registry segment, because under-allocating the registry would be worse than the waste it replaced, and it asserts against a default uncaught exception handler rather than against the iteration result, since a dead worker does not change the result. Populating 2^31 entities is not feasible, but the ids only have to be sparse, so both use the package-private GigaMap.Default constructor that seeds the id counter; setupAddingState then allocates just the one segment chain the seeded id points at. That also makes the 2^43 rejection reachable for about 26 MB, using exponents 20/20 to keep the seeded id at level3 index 8. Each case fails on the unfixed code with the specific wrong value, including the id collision.

+++ gigamap/gigamap/src/main/java/org/eclipse/store/gigamap/types/AbstractBitmapIterating.java
@@ -385,7 +385,8 @@ public abstract class AbstractBitmapIterating<E>
 	private void updateLevel3IterationState(final int level3Index)
 	{
 		this.currentLevel3Index = level3Index;
-		this.level2BaseId       = level3Index * BitmapLevel3.LEVEL_2_ID_COUNT;
+		// (long) cast mandatory: the int product overflows at level3 index 2^11, i.e. at entityId 2^31.
+		this.level2BaseId       = (long)level3Index * BitmapLevel3.LEVEL_2_ID_COUNT;
 		if(this.level2BaseId >= this.level2TrailingBaseId)
 		{
 			// when the trailing segment is reached, the level2 index bound is 1 beyond the upperId's level2 index.
